Scripts Updated
Added better docstrings and exceptions

+++ XRD/BrukerRAW/.ipynb_checkpoints/brukerrawconverter-checkpoint.py
@@ -4,6 +4,9 @@ import xylib
 
 #Function gets the metadata and writes it to a file. Used in main function
 def export_metadata(f, meta):
+    """
+    A function to extract metadata from files
+    """
     for i in range(meta.size()):
         key = meta.get_key(i)
         value = meta.get(key)
@@ -11,7 +14,19 @@ def export_metadata(f, meta):
 
 #Converts input file into a text file title "output_file.txt"
 def brukerrawconverter(input_file, output_file):
+    """
+    A function that takes an input file, extracts the data from the input, and writes it to the output file.
+    This function should work for all file types that the xylib library can handle, but it is primarily being used for Bruker RAW and Bruker XRD formats here.
 
+    args: first argument is input_file. Input_file is one of the supported file types. Second arguemnt is output_file. Output_file is a string that will be used for the name of the output file that is created in the function
+    return: does not return anything. Saves output_file as a text file
+    exceptions: will throw an exception if the input_file is not a file
+    """
+
+    #Check if input is a file
+    if os.path.isfile(input_file) is not true:
+        raise ValueError(“ERROR: bad input. Expected file”)
+    
     #Load input file and create output text file
     d = xylib.load_file(input_file)
     f = open(output_file,"w+")
